✨ Register use case
We've decoupled the logic responsible for database connection and user creation from the HTTP handling code. This separation enhances maintainability because if we ever need to create a user through a different mechanism—such as a CLI command, a background job, or an external integration—we can do so without relying on the HTTP route. This makes our codebase more modular, reusable, and adaptable to different use cases beyond just handling web requests.