⭐ Dependendy inversion - a.k.a. letter "D" from SOLID
If you look at the register-use-case.ts file, you'll see that an instance of the Prisma repository is created and then used to execute the create method.

Instead of instantiating a dependency (e.g., PrismaUsersRepository) directly inside the use case, we transformed RegisterUseCase from a function into a class. This allows us to pass the repository as a constructor parameter, ensuring that the use case works regardless of which repository is used.

This follows the Dependency Inversion Principle (DIP) from SOLID, ensuring that register-use-case.ts and PrismaUsersRepository remain independent of each other. As a result, if we want to use a different repository—for example, a local repository for testing—we don’t need to modify the use case file. Instead, we simply pass the new repository instance into the constructor of the RegisterUseCase class.