⭐ Test coverage
Coverage is a metric used to measure the amount of code being tested in an application. It generates a report of the lines of code that are being tested, and by analyzing them you can create the necessary tests.

Setup: Add and run the script in package.json > Install its dependency > Run it again > Add the coverage folder to .gitignore > Open the folder in the default file explorer > Open index.html to view the report.